+++ NEWS.md
@@ -1,8 +1,12 @@
 # ispv (development version)
 
-* added function for loading country-wide ISCO-based data
-* added function for loading country-wide NACE-based data
+* added function for loading country-wide ISCO-based data: `pv_cr_monthlypay_isco()`
+* added function for loading country-wide NACE-based data: `pv_cr_monthlypay_nace()`
+* added function for loading regional data on earnings by gender-age: `pv_reg_monthlypay_age_gender()`
+* added function for loading regional data on earnings by education: `pv_reg_monthlypay_education()`
+* added function for loading country-wide data on earnings by gender-age: `pv_cr_monthlypay_age_gender()`
+* added function for loading country-wide data on earnings by education: `pv_cr_monthlypay_education()`
 
 # ispv 0.1.0
 
-* Working functions for regional data
+* Working functions for regional data by profession (ISCO)